+++ src/govuk/components/cookie-banner/cookie-banner.story.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import fixtures from 'govuk-frontend/govuk/components/cookie-banner/fixtures.json';
+import { CookieBanner } from '.';
+import processExampleData from '../../../../utils/processExampleData';
+
+const stories = storiesOf('cookie-banner', module);
+
+for (const example of Object.values(
+  processExampleData(fixtures.fixtures.filter((fixture) => !fixture.hidden))
+)) {
+  stories.add(example.name, () => <CookieBanner {...example.options} />);
+}